@@ -36,15 +36,15 @@
   "homepage": "https://github.com/kuuurt13/ng-block-ui#readme",
   "dependencies": {},
   "devDependencies": {
-    "@angular/common": "^5.0.2",
-    "@angular/compiler": "^5.0.2",
-    "@angular/compiler-cli": "^5.0.2",
-    "@angular/core": "^5.0.2",
-    "@angular/forms": "^5.0.2",
-    "@angular/platform-browser": "^5.0.2",
-    "@angular/platform-browser-dynamic": "^5.0.2",
-    "@angular/platform-server": "^5.0.2",
-    "@angular/router": "^5.0.2",
+    "@angular/common": "6.0.7",
+    "@angular/compiler": "6.0.7",
+    "@angular/compiler-cli": "6.0.7",
+    "@angular/core": "6.0.7",
+    "@angular/forms": "6.0.7",
+    "@angular/http": "6.0.7",
+    "@angular/platform-browser": "6.0.7",
+    "@angular/platform-browser-dynamic": "6.0.7",
+    "@angular/router": "6.0.7",
     "@types/jasmine": "^2.5.43",
     "babel-polyfill": "^6.23.0",
     "es6-shim": "^0.35.1",
@@ -59,15 +59,15 @@
     "karma-webpack": "^1.7.0",
     "raw-loader": "^0.5.1",
     "reflect-metadata": "^0.1.3",
-    "rxjs": "^5.1.0",
+    "rxjs": "^6.2.1",
     "ts-loader": "^0.8.2",
     "tslint": "^3.13.0",
     "tslint-loader": "^2.1.5",
-    "typescript": "^2.4.2",
+    "typescript": "2.7.2",
     "typings": "^1.3.1",
     "watchman": "^0.1.8",
     "webpack": "^1.13.1",
     "webpack-dev-server": "^1.14.1",
-    "zone.js": "^0.8.4"
+    "zone.js": "^0.8.26"
   }
 }
