@@ -5,9 +5,8 @@ import {
   HttpEvent,
   HttpInterceptor
 } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
-import { _finally } from 'rxjs/operator/finally';
-
+import { Observable } from 'rxjs';
+import { finalize } from 'rxjs/operators';
 import { BlockUIService } from '../services/block-ui.service';
 import { BlockUIHttpSettings } from './block-ui-http-settings.service';
 import { BlockUIDefaultName } from '../constants/block-ui-default-name.constant';
@@ -22,21 +21,19 @@ export class BlockUIInterceptor implements HttpInterceptor {
   ) {}
 
   intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
-    const req = next.handle(request);
-    const response$ = req.subscribe.bind(req);
     let active = false;
 
-    req.subscribe = (...args) => {
-      if (this.shouldBlock(request)) {
-        this.blockUIService.start(BlockUIDefaultName);
-        active = true;
-      }
+    if (this.shouldBlock(request)) {
+      this.blockUIService.start(BlockUIDefaultName);
+      active = true;
+    }
 
-      return response$(...args);
-    };
-
-    return _finally
-      .call(req, () => active && this.blockUIService.stop(BlockUIDefaultName));
+    return next.handle(request)
+      .pipe(
+        finalize(() => {
+          active && this.blockUIService.stop(BlockUIDefaultName);
+        })
+      );
   }
 
   shouldBlock(request: HttpRequest<any>): boolean {
