@@ -1,12 +1,7 @@
 require('es6-shim');
 require('reflect-metadata');
 require('zone.js/dist/zone');
-require('zone.js/dist/long-stack-trace-zone');
-require('zone.js/dist/async-test');
-require('zone.js/dist/fake-async-test');
-require('zone.js/dist/sync-test');
-require('zone.js/dist/proxy');
-require('zone.js/dist/jasmine-patch');
+require('zone.js/dist/zone-testing');
 
 const browserTesting = require('@angular/platform-browser-dynamic/testing');
 const coreTesting = require('@angular/core/testing');
